requirements.txt
Flask
PyMuPDF
python-dotenv
Werkzeug
gunicorn
flask-limiter